+++ CHANGELOG.md
@@ -12,12 +12,13 @@
 - added `draw_offset` and `draw_offset_with_parallax_and_scroll` to `TextureInterface`;
 - grid skew and angle changes now mark the map as having unsaved changes;
 - improved the way the camera position is capped to the map size;
-- improved the reliability of the selection of map items.
+- improved the reliability of the selection of map items;
+- merged the brush and thing hardcoded properties in their respective `properties` hash maps.
 
 ### Fixes
 - Fixed overlapping actions in rotate tool;
 - fixed an issue where undoing changes past the last saved edit would not result in the map marked as in need of being saved when quitting the application;
-- fixed an issue where properties could all be converted to a String;
+- fixed an issue where properties could all be converted to a `String`;
 - fixes brush sprites being incorrectly rendered;
 - fixed entity tool crash when attempting to move textures but none were selected;
 - fixed one of the brushes created from a subtraction having no texture;